login: extract LoginClient and default HASH_SALT for tests
